+++ packages/vchart-types/types/component/label/util.d.ts
@@ -1,6 +1,7 @@
 import type { WaterfallSeries } from './../../series/waterfall/waterfall';
 import type { Datum } from '../../typings/common';
 import type { LabelItem, OverlapAttrs, Strategy } from '@visactor/vrender-components';
+import type { IGraphic, IText } from '@visactor/vrender-core';
 import type { ILabelInfo, ILabelSpec } from './interface';
 export declare const labelRuleMap: {
     rect: typeof barLabel;
@@ -15,6 +16,7 @@ export declare const labelRuleMap: {
     arc3d: typeof pieLabel;
     treemap: typeof treemapLabel;
     venn: typeof vennLabel;
+    boxPlot: typeof boxPlotLabel;
 };
 export declare function defaultLabelConfig(rule: string, labelInfo: ILabelInfo): any;
 export declare function textAttribute(labelInfo: ILabelInfo, datum: Datum, formatMethod?: ILabelSpec['formatMethod'], formatter?: ILabelSpec['formatter']): Partial<import("../../core").IComposedTextMarkSpec>;
@@ -47,7 +49,7 @@ export declare function pieLabel(labelInfo: ILabelInfo): {
 export declare function stackLabelX(datum2: Datum, series: WaterfallSeries, pos: string, offset: number): number;
 export declare function stackLabelY(datum2: Datum, series: WaterfallSeries, pos: string, offset: number): number;
 export declare function stackLabel(labelInfo: ILabelInfo, datumTransform?: (data: any) => any, attributeTransform?: (label: LabelItem, datum: Datum, att: any) => any): {
-    customLayoutFunc: (labels: LabelItem[]) => import("@visactor/vrender-core").IText[];
+    customLayoutFunc: (labels: LabelItem[]) => IText[];
     dataFilter: (labels: LabelItem[]) => LabelItem[];
     overlap: {
         strategy: any;
@@ -77,4 +79,7 @@ export declare function sankeyLabel(labelInfo: ILabelInfo): {
     smartInvert: boolean;
     offset: number;
     syncState: boolean;
+};
+export declare function boxPlotLabel(labelInfo: ILabelInfo): {
+    customLayoutFunc: (labels: LabelItem[], texts: IText[], getRelatedGraphic: (item: LabelItem) => IGraphic) => IText[];
 };